back/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.3
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
branca==0.4.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==2.2.4
Django==3.2.7
django-allauth==0.50.0
django-appconf==1.0.6
django-cors-headers==3.12.0
django-environ==0.11.2
django-extensions==3.1.5
django-imagekit==5.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-seed==0.3.1
djangorestframework==3.13.1
drf-yasg==1.20.0
entrypoints==0.4
exceptiongroup==1.2.1
executing==0.8.3
Faker==13.11.1
folium==0.12.1.post1
fonttools==4.29.1
frozenlist==1.4.1
h11==0.13.0
httpcore==1.0.5
httpx==0.27.0
idna==3.3
inflection==0.5.1
ipykernel==6.9.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itypes==1.2.0
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.3.1
jupyter_core==5.0.0
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.10
kiwisolver==1.3.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.7
mistune==0.8.4
mpmath==1.2.1
multidict==6.0.5
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.4
oauthlib==3.2.0
openai==0.28.0
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pilkit==3.0
pillow==10.3.0
platformdirs==4.2.2
prometheus-client==0.13.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.11.2
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.8.1
selenium==4.1.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.2.0
sympy==1.9
terminado==0.13.1
testpath==0.6.0
threadpoolctl==3.1.0
toposort==1.7
tornado==6.1
tqdm==4.66.4
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wsproto==1.1.0
xlrd==2.0.1
yarl==1.9.4
